integrated an ePharmacy system into our hospital management system.
Integrated an an ePharmacy system with a unique feature: doctors, while creating medical reports, can add products to a cart that's securely stored in the database. These products are categorized as "recommended" or "not recommended," and for the latter, the quantity can be modified. Upon report finalization, a unique code is generated and sent to the patient's email. This code allows the patient to purchase the products, ensuring security and adherence to regulations. Each code can only be used once and cannot be shared with other patients for added safety and control.